backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn==0.24.0
alembic==1.13.1
pytest==7.4.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
argon2-cffi==25.1.0
python-multipart==0.0.6
pydantic[email]==2.5.0
httpx==0.25.0
psycopg2-binary